third commit

+++ src/main/java/com/example/test3/service/StudentService.java
@@ -4,56 +4,49 @@ import com.example.test3.entity.Student;
 import com.example.test3.repository.StudentRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.Optional;
 
 @Service
 public class StudentService {
 
     @Autowired
-    private Student student;
-
-    @Autowired
     private StudentRepo studentRepo;
 
-
-
-    public void put(String name,int age){
-        Student newStudent=new Student();
-        newStudent.setName(name);
-        newStudent.setAge(age);
-
-        studentRepo.save(newStudent);
-
-
+    public void put(Student student, MultipartFile image) throws IOException {
+        if (image != null && !image.isEmpty()) {
+            student.setImageName(image.getOriginalFilename());
+            student.setImageType(image.getContentType());
+            student.setImageData(image.getBytes());
+        } else {
+            throw new RuntimeException("Image is required");
+        }
+        studentRepo.save(student);
     }
 
     public Student getStudentById(Long id) {
-        return studentRepo.findById(id).orElse(new Student());
+        return studentRepo.findById(id).orElseThrow(() -> new RuntimeException("Student not found"));
     }
 
     public void deleteStudent(Long id) {
-        studentRepo.deleteById(id);
+        if (studentRepo.existsById(id)) {
+            studentRepo.deleteById(id);
+        } else {
+            throw new RuntimeException("Student not found");
+        }
     }
 
-    public Student updateStudent(Long id, Student student) {
-
-        Optional<Student> student1=studentRepo.findById(id);
-
-        if(student1.isPresent()) {
-            Student student2 = student1.get();
-
-            student2.setName(student.getName());
-            student2.setAge(student.getAge());
-            return studentRepo.save(student2);
-
+    public void updateStudent(Long id, Student student) {
+        Optional<Student> existingStudent = studentRepo.findById(id);
+        if (existingStudent.isPresent()) {
+            Student studentToUpdate = existingStudent.get();
+            studentToUpdate.setName(student.getName());
+            studentToUpdate.setAge(student.getAge());
+            studentRepo.save(studentToUpdate);
+        } else {
+            throw new RuntimeException("Student not found");
         }
-
-        else {
-            return null;
-        }
-
-
-
     }
 }
